Notify the user when a the rift barrier unlocks.
Trigger a notification when the rift barrier's state goes from locked to
unlocked. This is implemented carefully so that the notification should
only trigger if the user is standing in front of the locked gate when it
unlocks (and shouldn't trigger on world hop or login).

Should be useful if you're waiting for a game in progress to end, but
you don't want to have to baby-sit the progress by peeking at the
portal.