Added alpha to color selectors in GuardiansOfTheRiftHelperConfig (#56)
* Added alpha to colors in GuardiansOfTheRiftHelperConfig

Added alpha to the color selectors. Tested in-game and found no new issues caused by the change.

* Added config request to essence pile outline colors in GuardiansOfTheRiftHelperOverlay

Added config request to essence pile outline colors. Also added essencePileColor() into the config.

* Adding alpha channel to color pickers GuardiansOfTheRiftHelperConfig.java

Added alpha channel to the color picker utility when ever it was used.

Added an option to pick the color of essence piles when the plugin recommends building one.